Add fix so Safari displays events on the flyer
The event start_time field returned from the Facebook Graph API is not
compliant with ISO 8601 - it is missing a colon in the timezone field.
Safari is (evidently) more strict about date string formats, so this
fixes the strings local to the flyer.

The events now display correctly in Safari on OS X and iOS.