Updated events.json
We forgot to remove the dates for the weeks we didn't have CoWorking Night during the holidays. Because of this it messed up our id numbers. This commit fixes that, removes the Big Spring Room from the list, and adds the Foundry as a list of places that can be used for workshops.